requirements.txt
# Azure Functions
azure-functions

# Azure Table Storage SDK
azure-data-tables

# For testing
pytest
